Velocity is now measured in ticks per second

+++ src/src/encoders.h
@@ -24,8 +24,8 @@ private:
 
     private:
         Encoder* encoder;
-        const int kp = 4, ki = 90;
-        int positionEstimate = 0, velocityEstimate = 0, velocityIntegrator = 0;
+        const double kp = 4.0, ki = 90.0;
+        double positionEstimate = 0, velocityEstimate = 0, velocityIntegrator = 0;
         unsigned int lastTime;
     };
